# server/public/openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "JCB Tracker API", "version": "2.0.0", "description": "JCB Tracker - İş Makineleri Takip Sistemi REST API" },
  "servers": [{ "url": "/api", "description": "API Sunucusu" }],
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } },
...
  },
  "paths": {
...
    "/config/{deviceId}": {
      "get": { "tags": ["Yapılandırma"], "summary": "Cihaz yapılandırması", "security": [{ "bearerAuth": [] }], "parameters": [{ "in": "path", "name": "deviceId", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Yapılandırma" } } },
      "put": { "tags": ["Yapılandırma"], "summary": "Cihaz yapılandırmasını güncelle", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Güncellendi" } } }
    },
...
    "/firmware/config/{deviceId}": {
      "get": { "tags": ["Firmware"], "summary": "ESP32 yapılandırma dosyası", "security": [{ "bearerAuth": [] }], "parameters": [{ "in": "path", "name": "deviceId", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Yapılandırma JSON" } } }
    },
...
  }
}